requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
streamlit
python-dotenv
requests
pandas
groq
langgraph
langchain
langchain-groq